+++ src/App.jsx
@@ -15,7 +15,8 @@ import TenantCriteria from './components/Tenant/TenantCriteria';
 import LandlordDashboard from './components/Landlord/LandlordDashboard';
 import ContactUs from './components/ContactUs';
 import NavBar from './components/NavBar';
-import LandlordContacts from './components/LandlordContacts';
+import LandlordContacts from './components/Landlord/LandlordContact';
+import Messages from './components/Landlord/LandlordContact/Messages';
 import Footer from './components/Footer';
 import GlobalStyle from './GlobalStyle';
 import SApp from './SApp';
@@ -39,10 +40,11 @@ export default function App() {
           <Route path="/tenant/dashboard" component={TenantDashboard} />
           <Route path="/landlord/dashboard" component={LandlordDashboard} />
           <Route path="/landlord/addannonce" component={AddAnnonce} />
+          <Route path="/landlord/messages/:id" component={Messages} />
           <Route path="/tenant/profil" component={TenantProfil} />
           <Route path="/tenant/criteria" component={TenantCriteria} />
           <Route path="/tenant/match" component={TenantMatch} />
-          <Route path="/contacts" component={LandlordContacts} />
+          <Route path="/landlord/contacts" component={LandlordContacts} />
           <Route
             path="/tenant/announcements-save"
             component={TenantAnnouncementsSave}
@@ -52,7 +54,6 @@ export default function App() {
             path="/tenant/housing-search"
             component={TenantHousingSearch}
           />
-
           <Route path="/contact-us" component={ContactUs} />
         </Switch>
       </UserInfos.Provider>
