+++ src/components/Landlord/LandlordContact/index.jsx
@@ -1,7 +1,38 @@
-export default function LandlordContact() {
+import { useState, useEffect } from 'react';
+import request from '../../../utilities/request';
+import SLandlordContacts from './style';
+import Contact from './Contact';
+
+export default function LandlordContacts() {
+  const [contacts, setContacts] = useState([]);
+  const [userInfos] = useState({ id: 1 });
+
+  useEffect(() => {
+    request({ method: 'get', url: `/contacts/${userInfos.id}` }).then(
+      ({ data }) => {
+        console.log(data);
+        setContacts(data);
+      }
+    );
+  }, []);
+
   return (
-    <>
-      <p>Contact</p>
-    </>
+    <div className="profil">
+      <SLandlordContacts>
+        <h2>Mes Contacts</h2>
+        <div className="icons">
+          {contacts &&
+            contacts.map((contact) => {
+              return (
+                <>
+                  {/*  <Link to="/landlord/messages/"> */}
+                  <Contact idSender={contact.id_sender} />
+                  {/*  </Link> */}
+                </>
+              );
+            })}
+        </div>
+      </SLandlordContacts>
+    </div>
   );
 }
